feat: add Spanish language support and translate content
This commit introduces Spanish language support to the documentation site by adding a new locale configuration in the VitePress config file.
It also includes the translation of all documentation pages into Spanish.

The changes include:
- Added Spanish locale configuration in `.vitepress/config.mts`
- Translated all documentation pages into Spanish
- Added Spanish navigation and sidebar configurations

This enhancement aims to make the Farcaster documentation more accessible to Spanish-speaking developers and users.